# swagger/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "M&M Car Rental Backend API Documentation",
        "description": "This is an API documentation for M&M Car Rental Backend API. You can find all the endpoints and their descriptions here.",
        "contact": {
            "email": "syahrezafistiferdian32@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "M&M Car Rental Backend Service"
        }
    ],
    "paths": {
...
        "/user/reservation/new": {
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Create a new car reservation",
                "description": "This endpoint creates a new reservation for a car to be rented by the user",
                "operationId": "b33d3e416a89516d78ea095666131997",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "car_id",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "car_id": {
                                        "description": "UUID of the car to be reserved",
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "start_date": {
                                        "description": "Start date of the reservation (YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "description": "End date of the reservation (YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 201
                                                },
                                                "is_success": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reservation created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "reservation_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "start_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-12-01"
                                                },
                                                "end_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-12-05"
                                                },
                                                "total_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1000000
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "car": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Toyota Avanza"
                                                        },
                                                        "year": {
                                                            "type": "integer",
                                                            "example": 2021
                                                        },
                                                        "jenis": {
                                                            "type": "string",
                                                            "example": "SUV"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data or bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 400
                                                },
                                                "is_success": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid data provided"
                                        },
                                        "data": {
                                            "description": "Validation errors",
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 500
                                                },
                                                "is_success": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to create reservation"
                                        },
                                        "data": {
                                            "type": "null",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "JWT Authorization header using the Bearer scheme.",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}